Refactor get_user_by_username, get_user_by_email; use across code base
