+++ lib/galaxy/visualization/genomes.py
@@ -11,6 +11,8 @@ from galaxy.exceptions import (
     ObjectNotFound,
     ReferenceDataError,
 )
+from galaxy.managers.users import get_user_by_username
+from galaxy.model import HistoryDatasetAssociation
 from galaxy.structured_app import StructuredApp
 from galaxy.util.bunch import Bunch
